fixes for IBM i - ambiguous types
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/10379


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@34285 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/decimfmt.cpp
@@ -1551,7 +1551,7 @@ DecimalFormat::_round(const DigitList &number, DigitList &adjustedNum, UBool& is
 
     if (fScale != 0) {
         DigitList ten;
-        ten.set(10);
+        ten.set((int32_t)10);
         if (fScale > 0) {
             for (int32_t i = fScale ; i > 0 ; i--) {
                 adjustedNum.mult(ten, status);
@@ -2196,7 +2196,7 @@ void DecimalFormat::parse(const UnicodeString& text,
 
         if (fScale != 0) {
             DigitList ten;
-            ten.set(10);
+            ten.set((int32_t)10);
             if (fScale > 0) {
                 for (int32_t i = fScale; i > 0; i--) {
                     UErrorCode ec = U_ZERO_ERROR;
